+++ project/mrjar/src/org/jpype/mrjar/A.java
@@ -1,0 +1,5 @@
+package org.jpype.mrjar;
+
+public class A
+{
+}