+++ native/java/org/jpype/pkg/JPypePackage.java
@@ -27,6 +27,7 @@ import java.util.ArrayList;
 import java.util.Map;
 import org.jpype.JPypeContext;
 import org.jpype.JPypeKeywords;
+import org.jpype.classloader.DynamicClassLoader;
 
 /**
  * Representation of a JPackage in Java.
@@ -42,12 +43,16 @@ public class JPypePackage
   // Name of the package
   final String pkg;
   // A mapping from Python names into Paths into the module/jar file system.
-  final Map<String, URI> contents;
+  Map<String, URI> contents;
+  int code;
+  private final DynamicClassLoader classLoader;
 
-  public JPypePackage(String pkg, Map<String, URI> contents)
+  public JPypePackage(String pkg)
   {
     this.pkg = pkg;
-    this.contents = contents;
+    this.contents = JPypePackageManager.getContentMap(pkg);
+    this.classLoader = ((DynamicClassLoader)(JPypeContext.getInstance().getClassLoader()));
+    this.code = classLoader.getCode();
   }
 
   /**
@@ -103,6 +108,7 @@ public class JPypePackage
    */
   public String[] getContents()
   {
+    checkCache();
     ArrayList<String> out = new ArrayList<>();
     for (String key : contents.keySet())
     {
@@ -215,6 +221,15 @@ public class JPypePackage
     {
       return false; // If anything goes wrong then it won't be considered a public class.
     }
+  }
+  
+  void checkCache()
+  {
+    int current = classLoader.getCode();
+    if (this.code == current)
+      return;
+    this.code = current;
+    this.contents = JPypePackageManager.getContentMap(pkg);
   }
 
 }