+++ doc/CHANGELOG.rst
@@ -8,6 +8,9 @@ Latest Changes:
 
   - Test build for Python 3.9
 
+  - Fixe bug when importing from multi-release jars.  The directory was
+    being truncated to only those classes in the overlay.
+
 - **1.1.2 - 2020-10-23**
 
   - Linux binaries are now stripped for size.